chore(fixture): rename fixtures
