+++ breadgood-server/src/test/java/com/bside/breadgood/fixtures/emoji/EmojiFixture.java
@@ -7,7 +7,7 @@ import com.bside.breadgood.ddd.emoji.domain.Emoji;
  * date : 2022/03/23
  * description :
  */
-public class EmojiFixtures {
+public class EmojiFixture {
     public static final Emoji 이모지1 =
             new Emoji(
                     1L,
